requirements.txt
sentencepiece>=0.1.90,<0.2.0
torch>=2.0
pytorch-crf>=0.7.2